\name{deadAKT_I}
\alias{deadAKT_I}
\docType{data}
\title{
Gene Response Characterization of AKT inhibitors: levels and factor structure.
}
\description{
This file shows the \code{name factors} and their \code{levels}, from two experiments (GSE18232, GSE18005) 
that could be used to determine the gene response to AKT inhibitors; data set taken from Krech et al. (2010).
}
\usage{data(deadAKT_I)}
\format{
This data frame contains \code{factors structure design}: \cr
\code{Gene name}, \code{Cell lines}, \code{Chemical compounds}, \code{Inhibitors groups} from \link{dataAKT_I} data.
5 chemical compounds which act as AKT ihibitors (group 1) were tested versus 4 chemical 
compounds which are not characterized as a direct AKT1 inhibitors (group 2) over three 
different cell lines (HCT116, HT29 and SW480).
}
\source{
Krech T, Thiede M, Hilgenberg E, Schadfer R, Jaijrchott K. (2010). Characterization of AKT independent effects of the 
synthetic AKT inhibitors SH-5 and SH-6 using an integrated approach combining transcriptomic profiling and signaling 
pathway perturbations. \emph{BMC Cancer. Jun 14;10:287}.
}
\references{
VARCOMPCI: A Package for Computation of Confidence Intervals for Variance Components of Mixed Models in R.
}
