\name{varcompci-internal}
\docType{methods}
\alias{initialize}
\alias{factorsf}
\alias{effectype}
\alias{subscriptss}
\alias{levelsf}
\alias{infoNLST}
\alias{combinposs}
\alias{EMSmatrix}
\alias{addRoFinmEMS}
\alias{addlev1inmEMS}
\alias{addRoF2inmEMS}
\alias{addlevinmEMS}
\alias{names2}
\alias{addRoF3inmEMS}
\alias{names3}
\alias{addRoF4inmEMS}
\alias{names4}
\alias{selectEMSout}
\alias{deletrows}
\alias{EMS}
\alias{EMSlF}
\alias{last_EMS}
\alias{covpar}
\alias{fvalues}
\alias{deletfixrow}
\alias{countnrm}
\alias{recAnova}
\alias{recAnovaII}
\alias{printaov}
\alias{printCI}
\alias{EMSwdesc}
\alias{printvariance}
\alias{printEMS}
\alias{printMeansq}
\alias{printanov}
\alias{printaic}
\alias{printci}
\alias{sasnoun}
\alias{sasauxnouns3}
\alias{sasauxnouns3b}
\alias{sasauxnouns4}
\alias{sasauxnouns4a}
\alias{sasauxnouns2aux}
\alias{sasauxnouns4b1}
\alias{sasauxnouns4b2}
\alias{sasauxnouns4b}
\alias{sasauxnouns5}
\alias{sasauxnouns5a}
\alias{sasauxnouns5b1}
\alias{sasauxnouns5b2}
\alias{sasauxnouns5b3}
\alias{sasauxnouns5b}
\alias{sasauxnouns5c1}
\alias{sasauxnouns5c2}
\alias{sasauxnouns5c31}
\alias{sasauxnouns5c32}
\alias{sasauxnouns5c3}
\alias{sasauxnouns5c}
\alias{ConfInt}
\alias{crearInforme}
\alias{.varcompciC.show}
\alias{.varcompciC.summary}
\alias{.EMSc.show}

\title{Internal varcompci objects and methods}
\description{
These are not to be called by the user.\cr
If you have a specific need or for more details on varcompci, feel free to contact the author (who, even if he is very lazy , will be very happy to answer and - why not? -start a new collaboration...)
}

\references{
 references
}
\section{Author(s)}{
 author
}


\seealso{
  Classes : \code{\linkS4class{varcompciC}}, \code{\linkS4class{EMSc}} \cr
  Methods : \code{\link{varcompci}}, \code{\link{prettyEMSf}} \cr
}
\keyword{internal}
